Update README.md
The Blood Bank App is an incredibly comprehensive platform for managing blood supply. It offers dedicated login pages for donors, admins, hospitals, and organizations, making it easy for each user to access the specific features they need. With a robust inventory list, donor list, and analytics option for tracking blood supply, this app is a must-have tool for anyone involved in managing blood donation. Whether you’re a donor looking to give blood, a hospital in need of a particular blood type, or an organization coordinating blood drives, the Blood Bank App has everything you need to streamline the process and ensure a safe and reliable blood supply